+++ src/state/gui/editor/ui/dialogs.rs
@@ -4,7 +4,7 @@ use rfd::FileDialog;
 
 use crate::{helper::concurrency::execution_queue::ExecutionQueue, state::scene::{components::material::TextureType, utilities::scene_utils::{load_sound, load_texture}}};
 
-pub fn load_texture_dialog(main_queue: Arc<RwLock<ExecutionQueue>>, texture_type: Option<TextureType>, scene_id: Option<u64>, material_id: Option<u64>, mipmapping: bool, max_tex_res: u32)
+pub fn load_texture_dialog(main_queue: Arc<RwLock<ExecutionQueue>>, texture_type: Option<TextureType>, scene_id: Option<u32>, material_id: Option<u32>, mipmapping: bool, max_tex_res: u32)
 {
     if let Some(path) = FileDialog::new().add_filter("Image", &["jpg", "png", "webp"]).set_directory("/").pick_file()
     {
@@ -26,7 +26,7 @@ pub fn load_texture_dialog(main_queue: Arc<RwLock<ExecutionQueue>>, texture_type
     }
 }
 
-pub fn load_sound_dialog(main_queue: Arc<RwLock<ExecutionQueue>>, sound_component_id: Option<u64>)
+pub fn load_sound_dialog(main_queue: Arc<RwLock<ExecutionQueue>>, sound_component_id: Option<u32>)
 {
     if let Some(path) = FileDialog::new().add_filter("Audio", &["ogg", "mp3", "wav", "flac"]).set_directory("/").pick_file()
     {
