+++ src/state/gui/editor/ui/lights.rs
@@ -6,7 +6,7 @@ use crate::{helper::{change_tracker::ChangeTracker, generic::cut_string_to_lengt
 
 use super::super::editor_state::{EditorState, SelectionType, SettingsPanel};
 
-pub fn build_light_list(editor_state: &mut EditorState, lights: &ChangeTracker<Vec<RefCell<ChangeTracker<LightItem>>>>, ui: &mut Ui, scene_id: u64)
+pub fn build_light_list(editor_state: &mut EditorState, lights: &ChangeTracker<Vec<RefCell<ChangeTracker<LightItem>>>>, ui: &mut Ui, scene_id: u32)
 {
     ui.with_layout(egui::Layout::top_down_justified(egui::Align::LEFT), |ui|
     {
@@ -59,7 +59,7 @@ pub fn create_light_settings(editor_state: &mut EditorState, state: &mut State, 
 {
     // no scene selected
     if editor_state.selected_scene_id.is_none() { return; }
-    let scene_id: u64 = editor_state.selected_scene_id.unwrap();
+    let scene_id: u32 = editor_state.selected_scene_id.unwrap();
 
     let (light_id, ..) = editor_state.get_object_ids();
