+++ src/state/scene/loader/wavefront.rs
@@ -20,9 +20,9 @@ pub fn get_texture_path(tex_path: &String, mtl_path: &str) -> String
     tex_path
 }
 
-pub fn load(path: &str, scene_id: u64, parent_node_id: Option<u64>, main_queue: ExecutionQueueItem, hide_root_node: bool, _reuse_materials: bool, _object_only: bool, create_mipmaps: bool, max_texture_resolution: u32) -> anyhow::Result<Vec<u64>>
+pub fn load(path: &str, scene_id: u32, parent_node_id: Option<u32>, main_queue: ExecutionQueueItem, hide_root_node: bool, _reuse_materials: bool, _object_only: bool, create_mipmaps: bool, max_texture_resolution: u32) -> anyhow::Result<Vec<u32>>
 {
-    let mut loaded_ids: Vec<u64> = vec![];
+    let mut loaded_ids: Vec<u32> = vec![];
 
     let resource_name = get_stem(path);
